pyproject.toml
[project]
name = "audio-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "gradio>=5.47.2",
    "langchain>=1.2.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "numpy>=2.3.3",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.2",
    "xgboost>=3.1.1",
]
